// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ActiveStatus = {
  ACTIVE: 'ACTIVE',
  SUSPEND: 'SUSPEND'
} as const

export type ActiveStatus = (typeof ActiveStatus)[keyof typeof ActiveStatus]


export const Role = {
  CUSTOMER: 'CUSTOMER',
  PROVIDER: 'PROVIDER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const GearStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type GearStatus = (typeof GearStatus)[keyof typeof GearStatus]


export const RentalOrderStatus = {
  PLACED: 'PLACED',
  CONFIRMED: 'CONFIRMED',
  PAID: 'PAID',
  PICKED_UP: 'PICKED_UP',
  RETURNED: 'RETURNED',
  CANCELLED: 'CANCELLED'
} as const

export type RentalOrderStatus = (typeof RentalOrderStatus)[keyof typeof RentalOrderStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]